fix: directoryをcursorのagentに頼んで修正してもらいました
