+++ lib/core/theme/app_theme.dart
@@ -1,0 +1,20 @@
+import 'package:flutter/material.dart';
+import 'package:google_fonts/google_fonts.dart';
+
+class AppTheme {
+  static ThemeData get light => ThemeData(
+        useMaterial3: true,
+        colorSchemeSeed: const Color(0xFF006E6D),
+        textTheme: GoogleFonts.nunitoTextTheme(),
+        cardTheme: CardTheme(
+          elevation: 0,
+          shape: RoundedRectangleBorder(
+            borderRadius: BorderRadius.circular(20),
+            side: BorderSide(
+              color: Colors.grey.shade200,
+              width: 1,
+            ),
+          ),
+        ),
+      );
+}
