+++ lib/features/settings/presentation/pages/settings_page.dart
@@ -1,0 +1,48 @@
+import 'package:flutter/material.dart';
+import 'package:go_router/go_router.dart';
+
+class SettingsPage extends StatelessWidget {
+  const SettingsPage({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        title: const Text('Settings'),
+        leading: IconButton(
+          icon: const Icon(Icons.arrow_back),
+          onPressed: () => context.pop(),
+        ),
+      ),
+      body: ListView(
+        padding: const EdgeInsets.all(16),
+        children: [
+          ListTile(
+            leading: const Icon(Icons.person),
+            title: const Text('Profile'),
+            subtitle: const Text('Set a nickname'),
+            onTap: () {},
+          ),
+          const Divider(),
+          SwitchListTile(
+            value: true,
+            onChanged: (_) {},
+            title: const Text('Daily reminder'),
+          ),
+          ListTile(
+            leading: const Icon(Icons.file_upload),
+            title: const Text('Export data (CSV)'),
+            onTap: () {},
+          ),
+          const Divider(),
+          ListTile(
+            leading: const Icon(Icons.delete, color: Colors.red),
+            title: const Text('Delete all data'),
+            textColor: Colors.red,
+            onTap: () {},
+          ),
+        ],
+      ),
+    );
+  }
+}
